+++ scripts/check-showcases.js
@@ -93,6 +93,7 @@ const numberedShowcases = [
   "30",
   "31",
   "32",
+  "33",
   "34",
 ];