+++ src/renderers/passes/index.ts
@@ -5,3 +5,4 @@ export * from "./PostProcessPass.js";
 export * from "./CascadedShadowPassGPU.js";
 export * from "./SpotShadowPassGPU.js";
 export * from "./DepthPrePassGPU.js";
+export * from "./HzbOcclusionPassGPU.js";